pkp/pkp-lib#7495 Move 'submissionSummaryModal' to own workflow folder + other renaming and moving around for better structure
