--- src/pages/dashboard/SubmissionSummaryModal/metaItems/BasicMetadata.vue
@@ -1,15 +1,0 @@
-<template>
-	<div>
-		<h3 class="text-lg-bold text-heading">{{ heading }}:</h3>
-		<p class="text-base-normal">{{ body }}</p>
-	</div>
-</template>
-<script setup>
-defineProps({
-	heading: {
-		type: String,
-		required: true,
-	},
-	body: {type: String, required: true},
-});
-</script>
