+++ src/pages/workflow/components/publication/WorkflowPublicationForm.vue
