+++ src/pages/workflow/components/publication/WorkflowPublicationJats.vue
