+++ src/pages/workflow/components/publication/WorkflowPublicationSectionJats.vue
@@ -145,6 +145,7 @@ export default {
 			},
 		},
 	},
+	emits: ['set'],
 	data() {
 		return {
 			isLoading: false,
